refactor(dashboard): extract pure derived selectors into useDashboardData
Move the genre charts, stats, activity, streak, rated/in-progress lists,
logged-title keys, and topGenre/lastPick selectors verbatim into
_hooks/use-dashboard-data.ts. All are deterministic functions of the
recommendations/libraryItems lists, so no side effects move; the data
fetch and handleMarkFinished mutation stay in the page. Behavior-
identical; page.tsx 2014->1867 LOC.
